feat: change wf
